+++ src/ti/job.c
@@ -9,6 +9,7 @@
 #include <ti/enum.inline.h>
 #include <ti/enums.inline.h>
 #include <ti/field.h>
+#include <ti/item.h>
 #include <ti/job.h>
 #include <ti/member.inline.h>
 #include <ti/method.h>
@@ -16,6 +17,7 @@
 #include <ti/names.h>
 #include <ti/procedure.h>
 #include <ti/procedures.h>
+#include <ti/prop.h>
 #include <ti/qbind.h>
 #include <ti/raw.inline.h>
 #include <ti/task.t.h>
@@ -94,6 +96,63 @@ static int job__set_add(ti_thing_t * thing, mp_unp_t * up)
         ti_val_unsafe_drop(v);
     }
 
+    return 0;
+}
+
+/*
+ * Returns 0 on success
+ * - for example: ``
+ */
+static int job__thing_clear(ti_thing_t * thing, mp_unp_t * UNUSED(up))
+{
+    if (ti_thing_is_dict(thing))
+        smap_clear(thing->items.smap, (smap_destroy_cb) ti_item_destroy);
+    else
+        vec_clear_cb(thing->items.vec, (vec_destroy_cb) ti_prop_destroy);
+
+    return 0;
+}
+
+/*
+ * Returns 0 on success
+ * - for example: `prop`
+ */
+static int job__arr_clear(ti_thing_t * thing, mp_unp_t * up)
+{
+    ti_val_t * val;
+    mp_obj_t mp_prop;
+
+    if (mp_next(up, &mp_prop) != MP_STR)
+    {
+        log_critical(
+                "job `clear` on array on "TI_THING_ID": "
+                "missing map, property or new_count",
+                thing->id);
+        return -1;
+    }
+
+    val = ti_thing_val_by_strn(thing, mp_prop.via.str.data, mp_prop.via.str.n);
+    if (!val)
+    {
+        log_critical(
+                "job `clear` on array on "TI_THING_ID": "
+                "missing property: `%.*s`",
+                thing->id,
+                mp_prop.via.str.n, mp_prop.via.str.data);
+        return -1;
+    }
+
+    if (!ti_val_is_list(val))
+    {
+        log_critical(
+                "job `clear` on array on "TI_THING_ID": "
+                "expecting a `"TI_VAL_LIST_S"`, got `%s`",
+                thing->id,
+                ti_val_str(val));
+        return -1;
+    }
+
+    vec_clear_cb(VARR(val), (vec_destroy_cb) ti_val_unsafe_gc_drop);
     return 0;
 }
 
@@ -2373,6 +2432,8 @@ int ti_job_run(ti_thing_t * thing, mp_unp_t * up)
     case TI_TASK_CLEAR_USERS:       break;
     case TI_TASK_TAKE_ACCESS:       break;
     case TI_TASK_ARR_REMOVE:        return job__arr_remove(thing, up);
+    case TI_TASK_THING_CLEAR:       return job__thing_clear(thing, up);
+    case TI_TASK_ARR_CLEAR:         return job__arr_clear(thing, up);
     case TI_TASK_SET_CLEAR:         return job__set_clear(thing, up);
     }